src/core/auth/schemas.py
from pydantic import BaseModel, EmailStr, Field
from src.core.enums import UserRole


class UserRegisterRequest(BaseModel):
    username: str = Field(min_length=3, max_length=50)
    email: EmailStr
    password: str = Field(min_length=6)
    full_name: str | None = None
    role: UserRole = UserRole.ADMIN


class UserLoginRequest(BaseModel):
    username: str
    password: str


class TokenResponse(BaseModel):
    access_token: str
    refresh_token: str
    token_type: str = "bearer"
    user_id: int
    username: str
    role: str


class RefreshTokenRequest(BaseModel):
    refresh_token: str


class UserResponse(BaseModel):
    id: int
    username: str
    email: str
    role: str
    full_name: str | None
    is_active: bool

    class Config:
        from_attributes = True