pyproject.toml
[project]
name = "test-project"
version = "0.1.0"
description = "Project for test"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.0",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "fastapi>=0.119.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[[tool.uv.index]]
url = "https://gitlab.pro-bit.si/api/v4/projects/88/packages/pypi/simple"
